+++ sdl2-sonic-drivers/src/hardware/opl/scummvm/mame/MameOPL2.hpp
@@ -26,10 +26,10 @@ namespace hardware::opl::scummvm::mame
         bool init() override;
         void reset() override;
 
-        void write(const uint32_t port, const uint8_t val) noexcept override;
+        void write(const uint32_t port, const uint16_t val) noexcept override;
         uint8_t read(const uint32_t port) noexcept override;
 
-        void writeReg(const uint32_t r, const uint8_t v) noexcept override;
+        void writeReg(const uint16_t r, const uint16_t v) noexcept override;
 
     protected:
         void generateSamples(int16_t* buffer, int length) noexcept override;
